+++ weblayer/browser/android/javatests/src/org/chromium/weblayer/test/InputTypesTest.java
@@ -1,14 +1,12 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2019 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package org.chromium.weblayer.test;
 
 import android.Manifest;
-import android.annotation.TargetApi;
 import android.app.Activity;
 import android.content.ClipData;
-import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
@@ -18,6 +16,7 @@ import android.os.Handler;
 import android.os.Parcelable;
 import android.provider.MediaStore;
 
+import androidx.annotation.RequiresApi;
 import androidx.core.app.ActivityCompat;
 import androidx.test.filters.SmallTest;
 
@@ -29,12 +28,13 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import org.chromium.base.Function;
+import org.chromium.base.ContextUtils;
+import org.chromium.base.test.util.ApplicationContextWrapper;
 import org.chromium.base.test.util.CallbackHelper;
 import org.chromium.base.test.util.Criteria;
 import org.chromium.base.test.util.CriteriaHelper;
-import org.chromium.base.test.util.InMemorySharedPreferencesContext;
 import org.chromium.base.test.util.MinAndroidSdkLevel;
+import org.chromium.base.test.util.UrlUtils;
 import org.chromium.content_public.browser.test.util.TestThreadUtils;
 import org.chromium.weblayer.shell.InstrumentationActivity;
 
@@ -51,6 +51,7 @@ public class InputTypesTest {
             new InstrumentationActivityTestRule();
 
     private File mTempFile;
+    private File mTestDir;
     private int mCameraPermission = PackageManager.PERMISSION_GRANTED;
 
     private class FileIntentInterceptor implements InstrumentationActivity.IntentInterceptor {
@@ -86,7 +87,7 @@ public class InputTypesTest {
 
     private FileIntentInterceptor mIntentInterceptor = new FileIntentInterceptor();
 
-    @TargetApi(Build.VERSION_CODES.M)
+    @RequiresApi(Build.VERSION_CODES.M)
     private class PermissionCompatDelegate implements ActivityCompat.PermissionCompatDelegate {
         public int mResult = PackageManager.PERMISSION_DENIED;
         private CallbackHelper mCallbackHelper = new CallbackHelper();
@@ -127,40 +128,45 @@ public class InputTypesTest {
 
     @Before
     public void setUp() throws Exception {
-        Bundle extras = new Bundle();
-        // We need to override the context with which to create WebLayer.
-        extras.putBoolean(InstrumentationActivity.EXTRA_CREATE_WEBLAYER, false);
-
-        Function<Context, Context> applicationContextBuilder = (baseContext) -> {
-            return new InMemorySharedPreferencesContext(baseContext) {
-                @Override
-                public int checkPermission(String permission, int pid, int uid) {
-                    if (permission.equals(Manifest.permission.CAMERA)) {
-                        return mCameraPermission;
+        ContextUtils.initApplicationContextForTests(
+                new ApplicationContextWrapper(ContextUtils.getApplicationContext()) {
+                    @Override
+                    public int checkPermission(String permission, int pid, int uid) {
+                        if (permission.equals(Manifest.permission.CAMERA)) {
+                            return mCameraPermission;
+                        }
+                        return super.checkPermission(permission, pid, uid);
                     }
-                    return getBaseContext().checkPermission(permission, pid, uid);
-                }
-            };
-        };
-        InstrumentationActivity.setActivityContextBuilder(applicationContextBuilder);
-
-        InstrumentationActivity activity = mActivityTestRule.launchShell(extras);
-        TestThreadUtils.runOnUiThreadBlocking(
-                () -> { activity.loadWebLayerSync(activity.getApplicationContext()); });
-        mActivityTestRule.navigateAndWait(mActivityTestRule.getTestDataURL("input_types.html"));
-        mTempFile = File.createTempFile("file", null);
-        activity.setIntentInterceptor(mIntentInterceptor);
+                });
         ActivityCompat.setPermissionCompatDelegate(mPermissionCompatDelegate);
 
+        mTestDir = new File(UrlUtils.getIsolatedTestFilePath("weblayer"));
+        if (!mTestDir.exists()) mTestDir.mkdir();
+        mTempFile = File.createTempFile("file", null, mTestDir);
         Intent response = new Intent();
         response.setData(Uri.fromFile(mTempFile));
         mIntentInterceptor.setResponse(Activity.RESULT_OK, response);
+
+        Bundle extras = new Bundle();
+        // We need to override the context with which to create WebLayer.
+        extras.putBoolean(InstrumentationActivity.EXTRA_CREATE_WEBLAYER, false);
+        InstrumentationActivity activity = mActivityTestRule.launchShell(extras);
+        activity.setIntentInterceptor(mIntentInterceptor);
+
+        TestThreadUtils.runOnUiThreadBlocking(
+                () -> activity.loadWebLayerSync(ContextUtils.getApplicationContext()));
+        mActivityTestRule.navigateAndWait(mActivityTestRule.getTestDataURL("input_types.html"));
     }
 
     @After
     public void tearDown() {
-        mTempFile.delete();
         ActivityCompat.setPermissionCompatDelegate(null);
+        if (mTempFile != null) {
+            mTempFile.delete();
+        }
+        if (mTestDir != null) {
+            mTestDir.delete();
+        }
     }
 
     @Test
@@ -259,7 +265,7 @@ public class InputTypesTest {
         Intent response = new Intent();
         ClipData clipData = ClipData.newUri(mActivityTestRule.getActivity().getContentResolver(),
                 "uris", Uri.fromFile(mTempFile));
-        File otherTempFile = File.createTempFile("file2", null);
+        File otherTempFile = File.createTempFile("file2", null, mTestDir);
         clipData.addItem(new ClipData.Item(Uri.fromFile(otherTempFile)));
         response.setClipData(clipData);
         mIntentInterceptor.setResponse(Activity.RESULT_OK, response);
